Conflicts:
	components/tools/OmeroWeb/omeroweb/webgateway/views.py

+++ components/server/src/ome/services/RenderingBean.java
@@ -997,7 +997,7 @@ public class RenderingBean implements RenderingEngine, Serializable {
         try {
             errorIfInvalidState();
             ChannelBinding[] cb = renderer.getChannelBindings();
-            return cb[w].getInputEnd().intValue();
+            return cb[w].getInputEnd().doubleValue();
         } finally {
             rwl.readLock().unlock();
         }
@@ -1015,7 +1015,7 @@ public class RenderingBean implements RenderingEngine, Serializable {
         try {
             errorIfInvalidState();
             ChannelBinding[] cb = renderer.getChannelBindings();
-            return cb[w].getInputStart().intValue();
+            return cb[w].getInputStart().doubleValue();
         } finally {
             rwl.readLock().unlock();
         }
